requirements.txt
mysql-connector-python==8.0.28
mysqlclient==2.1.0
fastapi==0.103.2
uvicorn==0.23.2
tenacity==8.0.1
